Shared form data
Closes #36
